Moved prometheus and autoclean job into subfolders
