+++ Dockerfile
@@ -7,6 +7,8 @@ COPY go.sum go.sum
 RUN go mod download
 
 COPY config/ config/
+COPY autoclean/ autoclean/
+COPY prometheus/ prometheus/
 COPY do_not_compress.go do_not_compress.go
 COPY main.go main.go
 RUN CGO_ENABLED=0 go build -o /main
